+++ src/test/java/com/dwarfeng/tmpstg/example/ConcurrentOperationExample.java
@@ -142,7 +142,7 @@ public class ConcurrentOperationExample {
         @Override
         public void run() {
             try {
-                int contentLength = RandomUtils.nextInt(contentLengthMin, contentLengthMax);
+                int contentLength = RandomUtils.insecure().randomInt(contentLengthMin, contentLengthMax);
                 byte[] content = ContentUtil.randomContent(contentLength);
                 String contentMd5 = ContentUtil.md5Checksum(content);
                 System.out.printf("任务 %d: 生成的内容的 MD5 值为 %s%n", index, contentMd5);
@@ -150,7 +150,7 @@ public class ConcurrentOperationExample {
                 try (OutputStream out = temporaryStorageHandler.openOutputStream(key, contentLength)) {
                     out.write(content);
                 }
-                long taskSleepDuration = RandomUtils.nextLong(taskSleepDurationMin, taskSleepDurationMax);
+                long taskSleepDuration = RandomUtils.insecure().randomLong(taskSleepDurationMin, taskSleepDurationMax);
                 System.out.printf("任务 %d: 休眠 %d 毫秒...%n", index, taskSleepDuration);
                 byte[] readContent = new byte[contentLength];
                 try (InputStream in = temporaryStorageHandler.openInputStream(key)) {